All working perfectly,just the readme to do and javadoc

+++ src/components/Sepia.java
@@ -25,9 +25,9 @@ public class Sepia extends Component{
 				int green_init = pixels[i][j][1];
 				int blue_init = pixels[i][j][2];
 				
-				pixels[i][j][0] = filterOverflow(0.393f, 0.769f, 0.189f, red_init, blue_init, green_init);
-				pixels[i][j][1] = filterOverflow(0.349f, 0.686f, 0.168f, red_init, blue_init, green_init);
-				pixels[i][j][2] = filterOverflow(0.272f, 0.534f, 0.131f, red_init, blue_init, green_init);
+				pixels[i][j][0] = filterOverflow(0.393, 0.769, 0.189, red_init, blue_init, green_init);
+				pixels[i][j][1] = filterOverflow(0.349, 0.686, 0.168, red_init, blue_init, green_init);
+				pixels[i][j][2] = filterOverflow(0.272, 0.534, 0.131, red_init, blue_init, green_init);
 				
 			}
 		}
@@ -35,7 +35,7 @@ public class Sepia extends Component{
 		return messageImage;
 	}
 
-	public int filterOverflow(float index1 ,float index2 ,float index3 ,int red_init , int blue_init , int green_init){
+	public int filterOverflow(double index1 ,double index2 ,double index3 ,int red_init , int blue_init , int green_init){
 		if((int)Math.round((red_init * index1 )+ (green_init * index2) + (blue_init * index3)) > 255)
 			return 255;
 		else return (int)Math.round((red_init * index1 )+ (green_init * index2) + (blue_init * index3));
